ISSUE-16 - Rename utils module to mapping
