+++ cargo-geiger/src/mapping/metadata.rs
